Rename widget to Switch
