+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "booleanslider",
-  "widgetName": "BooleanSlider",
+  "name": "switch",
+  "widgetName": "Switch",
   "version": "0.2.0",
   "description": "Toggle a boolean attribute",
   "scripts": {
