+++ src/package.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <package xmlns="http://www.mendix.com/package/1.0/">
-    <clientModule name="BooleanSlider" version="0.2.0" xmlns="http://www.mendix.com/clientModule/1.0/">
+    <clientModule name="Switch" version="0.2.0" xmlns="http://www.mendix.com/clientModule/1.0/">
         <widgetFiles>
-            <widgetFile path="BooleanSlider.xml"/>
+            <widgetFile path="Switch.xml"/>
         </widgetFiles>
         <files>
-            <file path="com/mendix/widget/custom/booleanslider"/>
+            <file path="com/mendix/widget/custom/switch"/>
         </files>
     </clientModule>
 </package>
